feat: FT4 protocol implementation and test suite
- Add FT4 decoding and encoding support
- Introduce tx_e2e test harness and golden WAVs for regression testing
- Add boot-time DRAM banner and CDC-FAIL watchdog
- Separate protocol configuration logic (feat: ENABLE_FT4)
- Remove MSVC-specific portability workarounds and obsolete tests

+++ main/station_types.h
@@ -33,7 +33,8 @@ enum class RadioType {
 };
 
 // One entry in the band list (name + frequency in kHz).
+// float to support FT4 frequencies with .5 kHz offsets (e.g. 40m FT4 = 7047.5 kHz).
 struct BandItem {
     const char* name;
-    int freq;  // kHz
+    float freq;  // kHz
 };
